file_management/requirements.txt
annotated-types==0.6.0
asgiref==3.8.0
attrs==23.2.0
cffi==1.16.0
contextlib2==21.6.0
cryptography==42.0.5
Django==5.0.3
django-filter==24.1
django-ninja==1.1.0
django-ninja-extra==0.20.4
django-ninja-jwt==5.3.0
djangorestframework==3.15.0
drf-spectacular==0.27.1
et-xmlfile==1.1.0
inflection==0.5.1
injector==0.21.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
Markdown==3.6
mysql-connector-python==8.3.0
mysqlclient==2.2.4
numpy==1.26.4
openpyxl==3.1.2
pandas==2.2.1
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
PyJWT==2.8.0
PyMySQL==1.1.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
rpds-py==0.18.0
setuptools==68.2.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
wheel==0.41.2
